+++ Sources/Protocols/AnyOptional.swift
@@ -1,0 +1,17 @@
+//
+//  AnyOptional.swift
+//  SwiftyUtils
+//
+//  Created by Tom Baranes on 25/04/2020.
+//  Copyright © 2020 Tom Baranes. All rights reserved.
+//
+
+import Foundation
+
+public protocol AnyOptional {
+    var isNil: Bool { get }
+}
+
+extension Optional: AnyOptional {
+    public var isNil: Bool { self == nil }
+}